.circleci: Add slash to end of s3 cp
This fixes the issue we had with the nightlies not being uploaded
properly, basically what was happening was that `aws s3 cp` doesn't
automatically distinguish between prefixes that are already
"directories" vs a single file with the same name.

This means that if you'd like to upload a file to a "directory" in S3
you need to suffix your destination with a slash.

Signed-off-by: Eli Uriegas <eliuriegas@fb.com>

ghstack-source-id: 776956819c8df453a28a32b636114f5be8b038a1
Pull Request resolved: https://github.com/pytorch/pytorch/pull/43792

Signed-off-by: Eli Uriegas <eliuriegas@fb.com>